allow for negative time again..
git-svn-id: file:///home/amuse/svn/amuse/trunk@5259 6e832d89-e7d1-4ad1-8ca1-873beecccf65

+++ src/amuse/community/huayno/src/evolve.h
@@ -211,5 +211,5 @@ struct sys join(struct sys s1,struct sys s2);
     ENDRUN("timestep too small: etime=%Le stime=%Le dt=%Le clevel=%u\n", etime, stime, dt, clevel); \
   } else { \
   if(etime == stime ||  dt==0 || clevel>=MAXLEVEL) \
-    ENDRUN("timestep too small: etime=%le stime=%le dt=%le clevel=%u\n", etime, stime, dt, clevel); \
+    ENDRUN("timestep too small: etime=%le stime=%le dt=%le clevel=%u\n", (double) etime, (double) stime, (double) dt, clevel); \
   }
